+++ tests/test_treeclass.py
@@ -20,7 +20,7 @@ from typing import Any
 import numpy.testing as npt
 import pytest
 
-from pytreeclass._src.backend import numpy as np
+from pytreeclass._src.backend import backend
 from pytreeclass._src.backend import tree_util as tu
 from pytreeclass._src.code_build import (
     autoinit,
@@ -32,6 +32,17 @@ from pytreeclass._src.code_build import (
 from pytreeclass._src.tree_base import TreeClass
 from pytreeclass._src.tree_mask import freeze
 from pytreeclass._src.tree_util import Partial, is_tree_equal
+
+if backend == "jax":
+    import jax.numpy as arraylib
+elif backend == "numpy":
+    import numpy as arraylib
+elif backend == "torch":
+    import torch as arraylib
+
+    arraylib.array = arraylib.tensor
+else:
+    raise ImportError("no backend installed")
 
 
 def test_fields():
@@ -133,8 +144,8 @@ def test_field_nondiff():
     class Test(TreeClass):
         def __init__(
             self,
-            a=freeze(np.array([1, 2, 3])),
-            b=freeze(np.array([4, 5, 6])),
+            a=freeze(arraylib.array([1, 2, 3])),
+            b=freeze(arraylib.array([4, 5, 6])),
         ):
             self.a = a
             self.b = b
@@ -144,12 +155,12 @@ def test_field_nondiff():
     assert tu.tree_flatten(test)[0] == []
 
     class Test(TreeClass):
-        def __init__(self, a=np.array([1, 2, 3]), b=np.array([4, 5, 6])):
+        def __init__(self, a=arraylib.array([1, 2, 3]), b=arraylib.array([4, 5, 6])):
             self.a = freeze(a)
             self.b = b
 
     test = Test()
-    npt.assert_allclose(tu.tree_flatten(test)[0][0], np.array([4, 5, 6]))
+    npt.assert_allclose(tu.tree_flatten(test)[0][0], arraylib.array([4, 5, 6]))
 
 
 def test_post_init():
@@ -273,12 +284,12 @@ def test_is_tree_equal():
 
     @autoinit
     class Test2(TreeClass):
-        a: np.ndarray = np.array([1, 2, 3])
+        a: Any = arraylib.array([1, 2, 3])
 
     assert is_tree_equal(Test1(), Test2()) is False
 
-    assert is_tree_equal(np.array([1, 2, 3]), np.array([1, 2, 3]))
-    assert is_tree_equal(np.array([1, 2, 3]), np.array([1, 3, 3])) is False
+    assert is_tree_equal(arraylib.array([1, 2, 3]), arraylib.array([1, 2, 3]))
+    assert is_tree_equal(arraylib.array([1, 2, 3]), arraylib.array([1, 3, 3])) is False
 
     @autoinit
     class Test3(TreeClass):
@@ -287,7 +298,7 @@ def test_is_tree_equal():
 
     assert is_tree_equal(Test1(), Test3()) is False
 
-    assert is_tree_equal(np.array([1, 2, 3]), 1) is False
+    assert is_tree_equal(arraylib.array([1, 2, 3]), 1) is False
 
 
 def test_mutable_field():